Title: 배열의 특정 위치 찾기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240203/배열의 특정 위치 찾기/find-specific-location-fo-array.c
@@ -18,6 +18,6 @@ int main() {
             cnt++;
         }
     }
-    avg = r / cnt;
+    avg = (double)r / cnt;
     printf("%d %.1lf", sum, avg);
 }
